Muelitas/requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-bootstrap-modal-forms==1.5.0
gunicorn==20.0.4
html5lib==1.0.1
Pillow==6.2.1
psycopg2==2.8.4
PyPDF2==1.26.0
python-decouple==3.3
pytz==2019.3
reportlab==3.5.32
six==1.13.0
sqlparse==0.3.0
webencodings==0.5.1
whitenoise==4.1.4
xhtml2pdf==0.2.3
